news_scraper/requirements.txt
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
idna==3.4
joblib==1.2.0
lxml==4.9.2
nltk==3.8.1
pymongo==4.3.3
python-dotenv==1.0.0
regex==2023.3.23
requests==2.28.2
soupsieve==2.4.1
tqdm==4.65.0
urllib3==1.26.15
